langchain_pilot/pyproject.toml
[project]
name = "langchain-pilot"
version = "0.1.0"
description = "Sid's experiments with LangChain"
authors = [
    {name = "Sid Chavan",email = "contact@telosium.com"}
]
readme = "README.md"
requires-python = "<3.14,>=3.9"
dependencies = [
    "langchain (>=0.3.15,<0.4.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "langchain-openai (>=0.3.1,<0.4.0)",
    "langchain-community (>=0.3.15,<0.4.0)",
    "langchain-google-genai (>=2.0.9,<3.0.0)",
    "langchain-google-firestore (>=0.5.0,<0.6.0)",
    "firestore (>=0.0.8,<0.0.9)",
    "chromadb (>=0.6.3,<0.7.0)",
    "tiktoken (>=0.8.0,<0.9.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "firecrawl-py (>=1.10.2,<2.0.0)",
    "langchainhub (>=0.1.21,<0.2.0)",
    "wikipedia (>=1.4.0,<2.0.0)",
    "tavily-python (>=0.5.0,<0.6.0)",
    "langchain-anthropic (>=0.3.3,<0.4.0)",
    "youtube-transcript-api (>=0.6.3,<0.7.0)",
    "pytube (>=15.0.0,<16.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
